requirements.txt
Django==2.0
django-addanother==2.0.0
django-autocomplete-light==3.3.2
django-contrib-comments==1.9.0
django-countries==5.3.2
django-crispy-forms==1.7.2
django-filter==2.0.0
django-fsm==2.6.0
django-import-export==1.2.0
django-money==0.14.4
django-reversion==2.0.10
django-tables2==2.0.3
django-tellme==0.6.5
django-admin-numeric-filter==0.1.3
django-recaptcha==2.0.2
